Create an interface for weight conversion
- Instead of having one module per conversion, use one module, and individual classes will handle the logic of the conversion
- Create an abstract class to have a common interface among conversions
