--- bioimageio/core/weight_converter/__init__.py
@@ -1,1 +1,0 @@
-"""coming soon"""
