--- bioimageio/core/weight_converter/keras/__init__.py
@@ -1,1 +1,0 @@
-# TODO: update keras weight converters
