--- bioimageio/core/weight_converter/torch/__init__.py
@@ -1,1 +1,0 @@
-# TODO: torch weight converters
