+++ tests/test_add_weights.py
